+++ app/components/yearbook/Navbar.tsx
