+++ app/components/yearbook/SearchBar.tsx
