+++ app/components/yearbook/house-icons/Action.tsx
