+++ app/components/yearbook/house-icons/Animation.tsx
