+++ app/components/yearbook/house-icons/Apocalypse.tsx
