+++ app/components/yearbook/house-icons/Comedy.tsx
