+++ app/components/yearbook/house-icons/Horror.tsx
