+++ app/components/yearbook/house-icons/Musical.tsx
