+++ app/components/yearbook/house-icons/Mystery.tsx
