+++ app/components/yearbook/house-icons/Space.tsx
