+++ app/components/yearbook/house-icons/Superhero.tsx
