+++ app/components/yearbook/house-icons/index.tsx
