+++ app/components/yearbook/index.tsx
