+++ app/components/yearbook/social-media-icons/index.tsx
