+++ app/components/yearbook/ts/interfaces/Student.ts
