+++ app/components/yearbook/ts/interfaces/index.ts
