+++ pages/_app.tsx
@@ -1,8 +1,7 @@
 import React from 'react'
-import 'tailwindcss/tailwind.css'
-import '../styles/globals.css'
+import '../app/styles/globals.css'
 import type { AppProps } from 'next/app'
-import Layout from '../components/layout/Layout'
+import Layout from '../app/layout/Layout'
 
 /**
  * 描述
