+++ pages/yearbook/index.tsx
@@ -7,9 +7,9 @@ import {
 	YearbookFilterDropdown,
 	YearbookSortDropdown,
 	YearbookPagination,
-} from '../../components/yearbook'
-import { students } from '../../utils/students'
-import { Student } from '../../components/yearbook/ts/interfaces'
+} from '../../app/components/yearbook'
+import { students } from '../../app/utils/students'
+import { Student } from '../../app/components/yearbook/ts/interfaces'
 import styled from 'styled-components'
 
 const GradientBackgroundRed = styled.div`
@@ -43,30 +43,30 @@ const CharactersPage = () => {
 			<GradientBackgroundRed />
 			<YearbookNavbar />
 			<YearbookContainer>
-				<h1 className='flex justify-center mt-16 md:text-6xl text-4xl text-center font-bold text-white bg-transparent'>
+				<h1 className='flex justify-center mt-16 text-4xl font-bold text-center text-white bg-transparent md:text-6xl'>
 					Our Characters
 				</h1>
-				<p className='flex justify-center mb-20 mt-8 text-white text-center bg-transparent'>
+				<p className='flex justify-center mt-8 mb-20 text-center text-white bg-transparent'>
 					Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer nunc,
 					leo eros, tellus sit eget auctor.
 				</p>
 				<form
-					className='bg-transparent z-10'
+					className='z-10 bg-transparent'
 					action='submit'
 					method='POST'
 					onSubmit={event => event.preventDefault()}>
 					<YearbookSearchBar />
-					<div className='z-10 bg-transparent flex flex-col md:flex-row justify-center md:space-x-20 md:space-y-0 space-y-5 mx-auto mb-8 mt-14'>
+					<div className='z-10 flex flex-col justify-center mx-auto mb-8 space-y-5 bg-transparent md:flex-row md:space-x-20 md:space-y-0 mt-14'>
 						<YearbookSortDropdown />
 						<YearbookFilterDropdown />
 					</div>
 				</form>
 			</YearbookContainer>
-			<div className='flex flex-wrap justify-center mx-auto container md:px-0 px-4 md:mt-20 bg-transparent w-full'>
+			<div className='container flex flex-wrap justify-center w-full px-4 mx-auto bg-transparent md:px-0 md:mt-20'>
 				{students.map((student: Student) => (
 					<div
 						key={student.id}
-						className='lg:w-3/12 md:w-4/12 sm:w-6/12 w-full bg-transparent'>
+						className='w-full bg-transparent lg:w-3/12 md:w-4/12 sm:w-6/12'>
 						<YearbookProfileCard
 							key={student.id}
 							name={student.name}
